refactor(Function): Validator 기능 분리 및 Function class 이름 변경
기능 별 정적 메소드 분리 (Validator), 이름 Utils로 기타 정적 메소드 집합 클래스 변경
